Refactored JOI Validation to Separate File and Implemented Middleware
- Moved JOI validation schema to a new file named `schemas.js` for better modularity and organization.

- Created a new middleware function `validateCampground` to handle validation of campground data using JOI.

- Updated the POST and PUT routes for campgrounds to use the `validateCampground` middleware.

- This refactor improves code readability and reusability, ensuring that all campground data is validated consistently across different routes.
