+++ app.js
@@ -6,7 +6,7 @@ const port = process.env.PORT || 3000;
 const Campground = require("./models/campground");
 const ExpressError = require("./utilities/ExpressError");
 const catchAsync = require("./utilities/catchAsync");
-const Joi = require("joi");
+const { campgroundSchema } = require("./schemas")
 
 const methodOverride = require("method-override");
 const ejsMate = require("ejs-mate");
@@ -33,109 +33,84 @@ async function main() {
 	console.log("Database is successfully connected.");
 }
 
+const validateCampground = (req, res, next) => {
+
+	// use the JOI .validate method to get any errors
+	// take the 'error' key from the resulting object after .validate(req.body)
+	const { error } = campgroundSchema.validate(req.body);
+	if (error) {
+		// use map to iterate over each object in the 'details' array and extract the error message
+		const msg = error.details.map((element) => element.message).join(",");
+		// console.log(msg);
+		throw new ExpressError(msg, 400);
+	} else {
+		next();
+	}
+};
+
 // ROUTES
 app.get("/", (req, res) => {
 	res.render("home");
 });
 
-app.get(
-	"/campgrounds",
-	catchAsync(async (req, res) => {
-		const allCampgrounds = await Campground.find({});
-		res.render(`campgrounds/index`, { allCampgrounds });
-	})
+app.get("/campgrounds", catchAsync(async (req, res) => {
+	const allCampgrounds = await Campground.find({});
+	res.render(`campgrounds/index`, { allCampgrounds });
+})
 );
 
 app.get("/campgrounds/new", (req, res) => {
 	res.render("campgrounds/newCampground");
 });
 
-app.post(
-	"/campgrounds",
-	catchAsync(async (req, res, next) => {
-		// if (!req.body.campground) {
-		// 	throw new ExpressError("Invalid Campground Data", 400);
-		// }
+// Middleware to Validate a new campground using JOI
 
-		// not a Mongoose schema; this uses Joi to validate a campground object before involving mongoose and saving to the database
-		const campgroundSchema = Joi.object({
-			campground: Joi.object({
-				title: Joi.string().required(),
-				price: Joi.number().required().min(0),
-			}).required(),
-			image: Joi.string().required(),
-			location: Joi.string().required(),
-			description: Joi.string().required(),
-		});
+app.post("/campgrounds", validateCampground, catchAsync(async (req, res, next) => {
+	// if (!req.body.campground) {
+	// 	throw new ExpressError("Invalid Campground Data", 400);
+	// }
 
-		// const result = campgroundSchema.validate(req.body);
-		// console.log(result);
-		// console.log(result.error);
+	const campground = new Campground(req.body.campground)
+	await campground.save();
+	res.redirect(`/campgrounds/${campground._id}`);
+})
+);
 
-		// take the 'error' key from the resulting object after .validate(req.body)
-		const { error } = campgroundSchema.validate(req.body);
-		if (error) {
-			// use map to iterate over each object in the 'details' array and extract the error message
-			const msg = error.details.map((element) => element.message).join(",");
-			console.log(msg);
-			throw new ExpressError(msg, 400);
+app.get("/campgrounds/:id/edit", catchAsync(async (req, res) => {
+	const { id } = req.params;
+	const campground = await Campground.findById(id);
+	res.render("campgrounds/editCampground", { campground });
+})
+);
+
+app.put("/campgrounds/:id", validateCampground, catchAsync(async (req, res) => {
+	const { id } = req.params;
+	const updatedCampground = await Campground.findByIdAndUpdate(
+		id,
+		{ ...req.body.campground },
+		{
+			runValidators: true,
+			new: true,
 		}
+	);
 
-		// console.log(result.error.details[0].message);
-
-		// if (result.error) {
-		// 	throw new ExpressError(result.error.details[0].message, 400);
-		// }
-
-		const campground = new Campground(req.body.campground);
-		await campground.save();
-		res.redirect(`/campgrounds/${campground._id}`);
-	})
+	console.log(updatedCampground);
+	res.redirect(`/campgrounds/${updatedCampground._id}`);
+})
 );
 
-app.get(
-	"/campgrounds/:id/edit",
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const campground = await Campground.findById(id);
-		res.render("campgrounds/editCampground", { campground });
-	})
+app.delete("/campgrounds/:id", catchAsync(async (req, res) => {
+	const { id } = req.params;
+	const deleteCampground = await Campground.findByIdAndDelete(id);
+	res.redirect("/campgrounds");
+})
 );
 
-app.put(
-	"/campgrounds/:id",
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const updatedCampground = await Campground.findByIdAndUpdate(
-			id,
-			{ ...req.body.campground },
-			{
-				runValidators: true,
-				new: true,
-			}
-		);
-
-		console.log(updatedCampground);
-		res.redirect(`/campgrounds/${updatedCampground._id}`);
-	})
-);
-
-app.delete(
-	"/campgrounds/:id",
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const deleteCampground = await Campground.findByIdAndDelete(id);
-		res.redirect("/campgrounds");
-	})
-);
-
-app.get(
-	"/campgrounds/:id",
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const campground = await Campground.findById(id);
-		res.render("campgrounds/showDetails", { campground });
-	})
+app.get("/campgrounds/:id", catchAsync(async (req, res) => {
+	const { id } = req.params;
+	const campground = await Campground.findById(id);
+	res.render("campgrounds/showDetails", { campground });
+})
 );
 
 // Catch all for routes that don't match above routes
